Fixed point upscaling algorithm updated. Displaying is not correct - need to be fixed.

+++ Software/InfraEye/module_LCD.h
@@ -6,7 +6,7 @@ void LCD_drawImage(uint16_t *colorArray, uint16_t columns, uint16_t rows, uint16
 void LCD_Convert(float *frameTemperature, uint16_t *frameColor, uint16_t frameSize,
   float minTemperature, float maxTemperature, uint8_t subPage);
 void LCD_Convert_u16(uint16_t *frameTemperature, uint16_t *frameColor, uint16_t frameSize,
-  float minTemperature, float maxTemperature, uint8_t subPage);
+  uint16_t minTemperature, uint16_t maxTemperature, uint8_t subPage);
 void LCD_FillRect(uint16_t x, uint16_t y, uint16_t width, uint16_t height, uint16_t color);
